feat: Add meta tags for SEO and social sharing
Add a meta description, Open Graph (OGP), and Twitter Card tags to `index.html`. These changes improve how the site appears in search engine results and provide rich previews when the URL is shared on social media.

A new `ogp.png` image is also added to be used as the preview image for these social cards.