refactor: Scroll to grid only when filtering by streamer
Previously, the page would scroll to the archive grid every time the filters were updated, including when toggling tags. This created a jarring user experience.

This change moves the `scrollIntoView` logic from the general `filterArchives` method to the more specific `filterByStreamer` method. Now, the page only scrolls to the top of the archive list when a user applies a streamer filter, resulting in a smoother interaction when adjusting tags.