feat(filter): Allow deselecting streamer filter by clicking again
Updates the streamer filter behavior to be more intuitive.

Previously, if a single streamer was selected, clicking their filter button again had no effect. This change modifies the logic so that a second click deselects the streamer and resets the filter to show all streamers.

This is implemented by calling `selectAllStreamers()` instead of returning early when an active filter is clicked.